chunked: fix build on other platforms
fix package build on !linux platforms
